+++ index.html
@@ -2,20 +2,10 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" href="/public/favicon.ico" type="image/x-icon" />
-    <link
-      rel="icon"
-      type="image/png"
-      href="/public/favicon-32.png"
-      sizes="32x32"
-    />
-    <link
-      rel="icon"
-      type="image/png"
-      href="/public/favicon-16.png"
-      sizes="16x16"
-    />
-    <link rel="apple-touch-icon" href="/public/apple-touch-icon.png" />
+    <link rel="icon" href="/favicon.ico" type="image/x-icon" />
+    <link rel="icon" type="image/png" href="/favicon-32.png" sizes="32x32" />
+    <link rel="icon" type="image/png" href="/favicon-16.png" sizes="16x16" />
+    <link rel="apple-touch-icon" href="/apple-touch-icon.png" />
 
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Portfolio</title>
