+++ src/App.jsx
@@ -1,16 +1,16 @@
-import Hero from "./components/Hero";
-import Tech from "./components/Tech";
-import Projects from "./components/Projects";
-import Skills from "./components/Skills";
-import Blog from "./components/Blog";
-import Contact from "./components/Contact";
+import Hero from "./components/Hero/Hero.jsx";
+import Tech from "./components/Tech/Tech.jsx";
+import Projects from "./components/Projects/Projects.jsx";
+import Skills from "./components/Skills/Skills.jsx";
+import Blog from "./components/Blog/Blog.jsx";
+import Contact from "./components/Contact/Contact.jsx";
 
 export const App = () => {
   return (
     <>
       <Hero
-        logo="../public/logo192.png"
-        img="../public/hero-img.png"
+        logo="/logo192.png"
+        img="/hero-img.png"
         name="Christina Baldwin"
         title="A Frontend Developer with a Background in Sustainability"
       />
