+++ src/components/Blog/BlogArticle.css
@@ -1,0 +1,9 @@
+.article {
+  display: flex;
+  gap: 3rem;
+}
+
+.article img {
+  max-width: 50%;
+  border-radius: 5px;
+}
