+++ src/components/Contact/Contact.css
@@ -1,0 +1,3 @@
+.contact {
+  text-align: center;
+}
