+++ src/components/Contact/Contact.jsx
@@ -1,3 +1,5 @@
+import "./Contact.css";
+
 const Contact = () => {
   return (
     <div className="contact">
