+++ src/components/Hero/Hero.css
@@ -1,0 +1,7 @@
+.hero {
+  text-align: center;
+}
+
+.logo {
+  width: 3%;
+}
