+++ src/components/Hero/HeroImages.css
@@ -1,0 +1,33 @@
+.hero-imgs {
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  aspect-ratio: 3 / 1;
+  height: 45rem;
+}
+
+.hero-img {
+  position: absolute;
+  width: 30%;
+  max-width: 300px;
+  transition: transform 0.3s ease, z-index 0.3s ease;
+  border-radius: 8px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.2);
+}
+
+.img-1 {
+  transform: rotate(-10deg) translateX(-100%);
+  z-index: 1;
+}
+
+.img-3 {
+  transform: rotate(10deg) translateX(100%);
+  z-index: 1;
+}
+
+.img-2 {
+  transform: rotate(0deg);
+  z-index: 2;
+}
