+++ src/components/Hero/HeroImages.jsx
@@ -1,0 +1,13 @@
+import "./HeroImages.css";
+
+const HeroImages = () => {
+  return (
+    <div className="hero-imgs">
+      <img src="/hero-img-1.jpg" className="hero-img img-1" />
+      <img src="/hero-img-2.png" className="hero-img img-2" />
+      <img src="/hero-img-3.jpg" className="hero-img img-3" />
+    </div>
+  );
+};
+
+export default HeroImages;
