--- src/components/HeroImages.jsx
@@ -1,11 +1,0 @@
-const HeroImages = () => {
-  return (
-    <div className="hero-imgs">
-      <img src="../public/hero-img-1.jpg" className="hero-img img-1" />
-      <img src="../public/hero-img-2.png" className="hero-img img-2" />
-      <img src="../public/hero-img-3.jpg" className="hero-img img-3" />
-    </div>
-  );
-};
-
-export default HeroImages;
