+++ src/components/Projects/Projects.jsx
@@ -1,4 +1,4 @@
-import projectsData from "../data/projects.json";
+import projectsData from "../../data/projects.json";
 import Project from "./Project";
 
 const Projects = () => {
