+++ src/index.css
@@ -1,84 +1,37 @@
 body {
   background: black;
   color: white;
+  line-height: 1.5;
 }
 
-/* HERO */
-.hero {
+h1 {
+  font-size: 5rem;
+}
+
+h2 {
+  font-size: 3rem;
+}
+
+h3 {
+  font-size: 2rem;
+}
+
+p {
+  font-size: 1.3rem;
+}
+
+button {
+  background-color: #4ecdc4;
+  border: none;
+  padding: 0.5rem 0.7rem;
+  font-size: 1rem;
+  display: inline-block;
+}
+
+.text-center {
   text-align: center;
 }
 
-.logo {
-  width: 3%;
-}
-
-.hero-imgs {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  aspect-ratio: 3 / 1;
-  height: 45rem;
-}
-
-.hero-img {
-  position: absolute;
-  width: 30%;
-  max-width: 300px;
-  transition: transform 0.3s ease, z-index 0.3s ease;
-  border-radius: 8px;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.2);
-}
-
-.img-1 {
-  transform: rotate(-10deg) translateX(-100%);
-  z-index: 1;
-}
-
-.img-3 {
-  transform: rotate(10deg) translateX(100%);
-  z-index: 1;
-}
-
-.img-2 {
-  transform: rotate(0deg);
-  z-index: 2;
-}
-
-/* TECH */
-.tech {
-  text-align: center;
-}
-
-/* PROJECTS */
-.project ul {
-  list-style: none;
-  display: flex;
-}
-
-/* SKILLS */
-.skills-table {
-  display: flex;
-  gap: 2rem;
-}
-
-.skills-table ul {
-  list-style: none;
-}
-
-/* BLOG */
-.article {
-  display: flex;
-  gap: 3rem;
-}
-
-.article img {
-  max-width: 50%;
-  border-radius: 5px;
-}
-
-/* CONTACT */
-.contact {
-  text-align: center;
+.text-left {
+  text-align: left;
 }
